Add kafka repository tests

+++ src/main/scala/com/clicktale/pipeline/sessionsfinalizer/contracts/FinalizerService.scala
@@ -33,7 +33,7 @@ trait FinalizerService extends LazyLogging {
 }
 
 object FinalizerService {
-  case class Session(subsId: Int, pid: Int, sid: Int, createDate: ZonedDateTime)
+  case class Session(subsId: Int, pid: Int, sid: Long, createDate: ZonedDateTime)
   case class Metrics(audits: Audits, failed: Boolean, unprocessedCount: Int)
   case class Audits(skipped: Seq[Session], processed: Seq[Session], failed: Seq[Session])
